Use antlers for all livewire templates and crop images via glide. Closes #82
