+++ resources/views/web/cart/country-select.antlers.html
@@ -1,0 +1,9 @@
+{{ if countries | count > 1 }}
+    <aside class="text-right">{{ __('butik::web.ship_to') }}
+        <select name="country" wire:model="country" wire:change="updateCountry">
+            {{ countries }}
+            <option value="{{ code }}">{{ name }}</option>
+            {{ /countries }}
+        </select>
+    </aside>
+{{ /if }}
