+++ resources/views/web/components/cart-icon.antlers.html
@@ -1,12 +1,12 @@
 <div class="flex items-center">
-    <a class="relative block" href="{{ route('butik.cart') }}" rel="noopener, nofollow">
+    <a class="relative block" href="{{ route:butik.cart }}" rel="noopener, nofollow">
         <svg class="w-6" viewBox="0 0 21 22" fill="none" xmlns="http://www.w3.org/2000/svg">
             <path fill-rule="evenodd" clip-rule="evenodd" d="M4.19995 1L1.19995 5V19C1.19995 20.1046 2.09538 21 3.19995 21H17.1999C18.3045 21 19.1999 20.1046 19.1999 19V5L16.1999 1H4.19995Z" stroke="#2D3748" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
             <path d="M1.19995 5H19.1999" stroke="#2D3748" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
             <path d="M14.2 9C14.2 11.2091 12.4091 13 10.2 13C7.99081 13 6.19995 11.2091 6.19995 9" stroke="#2D3748" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
         </svg>
         <div class="absolute -m-2 bg-accent rounded-full w-5 h-5 text-xs border-2 border-white bottom-0 right-0 text-white flex justify-center items-center">
-            {{ $total }}
+            {{ total }}
         </div>
     </a>
 </div>
