+++ src/Http/Controllers/Web/CheckoutController.php
@@ -7,11 +7,14 @@ use Illuminate\Support\Facades\Session;
 use Jonassiewertsen\StatamicButik\Checkout\Customer;
 use Jonassiewertsen\StatamicButik\Checkout\Cart;
 use Jonassiewertsen\StatamicButik\Http\Models\Order;
+use Jonassiewertsen\StatamicButik\Http\Traits\MapCartItems;
 use Jonassiewertsen\StatamicButik\Shipping\Country;
 use Statamic\View\View as StatamicView;
 
 class CheckoutController extends Checkout
 {
+    use MapCartItems;
+
     public function delivery()
     {
         $customer = session()->has('butik.customer') ?
@@ -78,29 +81,5 @@ class CheckoutController extends Checkout
                 'customer' => (array) json_decode($order->customer),
                 'order'    => $order->toArray(),
             ]);
-    }
-
-    /**
-     * Antlers can't handle objects and collections very well.
-     * To make them play nice together, we will return an
-     * array with all needed informations for the
-     * checkout process.
-     */
-    private function mappedCartItems()
-    {
-        return Cart::get()->map(function ($item) {
-            return [
-                'available'      => $item->available,
-                'sellable'       => $item->sellable,
-                'availableStock' => $item->availableStock,
-                'slug'           => $item->slug,
-                'images'         => $item->images,
-                'name'           => $item->name,
-                'description'    => $item->description,
-                'single_price'   => $item->singlePrice(),
-                'total_price'    => $item->totalPrice(),
-                'quantity'       => $item->getQuantity(),
-            ];
-        });
     }
 }
