+++ src/Http/Livewire/CartList.php
@@ -3,11 +3,14 @@
 namespace Jonassiewertsen\StatamicButik\Http\Livewire;
 
 use Jonassiewertsen\StatamicButik\Checkout\Cart as ShoppingCart;
+use Jonassiewertsen\StatamicButik\Http\Traits\MapCartItems;
 use Jonassiewertsen\StatamicButik\Shipping\Country;
 use Livewire\Component;
 
 class CartList extends Component
 {
+    use MapCartItems;
+
     public $country;
 
     public function mount()
@@ -36,7 +39,7 @@ class CartList extends Component
     {
         return view('butik::web.cart.cart-list', [
             'countries'      => Country::list(),
-            'items'          => ShoppingCart::get(),
+            'items'          => $this->mappedCartItems(),
             'total_price'    => ShoppingCart::totalPrice(),
             'total_shipping' => ShoppingCart::totalShipping(),
         ]);
